pyproject.toml
[tool.poetry]
name = "regexclass"
version = "0.1.0"
description = ""
authors = ["Caio Machado <moliveiracaio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.6"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ghostscript = "^0.7"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
unidecode = "^1.3.8"
pypdf2 = "^3.0.1"
black = "^24.8.0"
isort = "^5.13.2"
schedule = "^1.2.2"
pytz = "^2024.2"
boto3 = "^1.35.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
